server/asset: move CompatibilityCheck to dex/networks
